requirements-dev.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
click==7.1.2
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.0
pandas==1.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
virtualenv==20.0.4
virtualenvwrapper-win==1.2.6
wrapt==1.11.2
xlrd==1.2.0
